requirements.txt
alembic==1.8.1
Babel==2.9.0
bcrypt==4.0.0
cffi==1.15.1
click==8.1.3
cryptography==37.0.4
fabric==2.7.1
Fabric3==1.14.post1
Flask==2.2.2
Flask-Migrate==3.1.0
Flask-Moment==1.0.4
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
invoke==1.7.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
packaging==21.3
paramiko==2.11.0
pathlib2==2.3.7.post1
psycopg2==2.9.3
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.6.0
pytz==2022.2.1
six==1.16.0
SQLAlchemy==1.4.40
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1
